docs: add mkdocs
