+++ crate/cli/src/tests/findex/mod.rs
@@ -1,3 +1,3 @@
-pub(crate) mod add_or_delete;
+pub(crate) mod index_or_delete;
 pub(crate) mod search;
 pub(crate) mod tests;
