+++ documentation/.gitignore
@@ -1,0 +1,3 @@
+*.pd*
+/site
+/venv
