+++ documentation/pandoc/cryptobib/changes.txt
@@ -1,0 +1,246 @@
+* 2021-10-07
+ToSC 2021 (Issues 1-3)
+
+* 2021-09-29
+TCHES 2021 (Issue 4)
+
+* 2021-08-24
+PoPETS 2021 (Numbers 2-4)
+
+* 2021-08-16
+CRYPTO 2021, EUROCRYPT 2021
+
+* 2020-07-15
+TCHES 2021 (Issue 3)
+
+* 2021-06-22
+ToSC 2020 (Issues 3-4)
+
+* 2021-05-21
+PKC 2021, RSA 2021
+
+* 2021-03-16
+SODA 2021
+
+* 2021-03-02
+ICISC 2020, TCHES 2021 (Issue 2), JC 2021 (Number 2)
+
+* 2021-02-08
+COSADE 2020, ITCS 2021
+
+* 2021-02-01
+FOCS 2020, JC 2021 (Number 1), PoPETS 2021 (Number 1)
+
+* 2020-12-18
+ASIACRYPT 2020, CANS 2020, ICICS 2020, INDOCRYPT 2020, ISC 2020, LATIN 2020, PROVSEC 2020, TCC 2020, WISA 2020
+
+* 2020-12-04
+TCHES 2021 (Issue 1)
+
+* 2020-12-01
+EPRINT 2020 (up to 2020/1496)
+
+* 2020-11-30
+CSF 2007-2020
+
+* 2020-11-09
+ASIACCS 2020, CCS 2020, PoPETS 2020 (Number 4)
+
+* 2020-09-30
+ACNS 2020 (Part II), PoPETS 2020
+
+* 2020-09-17
+ESORICS 2020, NDSS 2020, SCN 2020, TRUSTBUS 2020
+
+* 2020-09-03
+ACNS 2020 (Part I), IEEE SP 2020, IWSEC 2020, JCEng 2020 (Number 3), PODC 2020, TCHES 2020 (Issue 4)
+
+* 2020-08-20
+FC 2020, FCW 2020, USENIX 2020
+
+* 2020-08-13
+ACISP 2020, CRYPTO 2020
+
+* 2020-07-28
+ToSC 2019 (Issue 4), ToSC 2020 (Issues 1-2), TOSC 2020 (NIST Special Issue)
+
+* 2020-07-24
+AFRICACRYPT 2020, ICALP 2020 
+
+* 2020-07-02
+TCHES 2020 (Issue 3)
+
+* 2020-06-26
+ITC 2020, STOC 2020, JC 2020 (Number 3)
+
+* 2020-05-26
+FCW 2019, JCEng 2020 (Numbers 1-2), JC 2020 (Number 2)
+
+* 2020-05-05
+EUROCRYPT 2020
+
+* 2020-05-03
+PKC 2020
+
+* 2020-04-16
+ICICS 2019, PQCRYPTO 2020
+
+* 2020-03-20
+TCHES 2020 (Issue 2)
+
+* 2020-02-19
+ICISC 2019, RSA 2020, JC 2020 (Number 1)
+
+* 2020-02-07
+SAC 2019, WISA 2019, SODA 2020
+
+* 2020-01-09
+FOCS 2019, ITCS 2020
+
+* 2019-12-29
+EPRINT (up to 2019/1483)
+
+* 2019-12-17
+IMA 2019, INDOCRYPT 2019
+
+* 2019-11-27
+ASIACRYPT 2019, TCC 2019
+
+* 2019-11-20
+TCHES 2020 (Issue 1)
+
+* 2019-11-18
+ACM CCS 2019, CANS 2019, COSADE 2019, FC 2019, JCEng 2019 (Number 4)
+
+* 2019-10-09
+FC 2018, JC 2019 (Number 4), ToSC 2019 (Issue 3)
+
+* 2019-10-03
+PROVSEC 2019, IEEE SP 2019
+
+* 2019-09-27
+JCEng 2019, PoPETS 2019, USENIX 2019
+
+* 2019-09-19
+ASIACCS 2019, ESORICS 2019, LATINCRYPT 2019, PQCRYPTO 2019
+
+* 2019-09-06
+ISC 2019, TRUSTBUS 2019
+
+* 2019-08-16
+CRYPTO 2019, IWSEC 2019, LATINCRYPT 2017, PODC 2019, JC 2019 (Numbers 2-3), TCHES 2019 (Issue 4)
+
+* 2019-07-18
+ICALP 2019
+
+* 2019-07-04
+ACISP 2019, AFRICACRYPT 2019
+
+* 2019-06-24
+STOC 2019, ToSC 2019 (Issue 2)
+
+* 2019-06-03
+ACNS 2019
+
+* 2019-05-21
+NIST PQC STANDARDIZATION ROUND 2
+
+* 2019-05-15
+EUROCRYPT 2019
+
+* 2019-05-09
+WISA 2018, NDSS 2019, TCHES 2019 (Issue 3)
+
+* 2019-04-10
+JC 2019 (Number 1), PKC 2019
+
+* 2019-03-07
+TCHES 2019 (Issue 2), ToSC 2019 (Issue 1)
+
+* 2019-03-04
+RSA 2019
+
+* 2019-02-27
+COSADE 2012-2018, FSW 2018, JCEng 2011-2018
+
+* 2019-02-22
+USENIX 1993-2018
+
+* 2019-02-20
+PETS 2002-2014, PoPETS 2015-2018
+
+* 2019-01-29
+ICISC 2018, SAC 2018
+
+* 2019-01-10
+ITCS 2019, SODA 2019
+
+* 2019-01-09
+JC 2019 (Number 1)
+
+* 2019-01-03
+FOCS 2018, INDOCRYPT 2018
+
+* 2018-12-25
+ToSC 2018 (Issue 4)
+
+* 2018-11-29
+CANS 2017, ICISC 2017
+
+* 2018-11-28
+ASIACRYPT 2018
+
+* 2018-11-20
+TCC 2018
+
+* 2018-11-15
+TCHES 2019 (Issue 1)
+
+* 2018-11-05
+TCHES 2018 (Issue 3)
+
+* 2018-10-31
+ACM CCS 2018, ICICS 2018
+
+* 2018-10-15
+CANS 2018, PROVSEC 2018
+
+* 2018-10-03
+JC 2018 (Number 4)
+
+* 2018-09-16
+PQCRYPTO 2008-2018
+
+* 2018-09-15
+FOCS 1996 (was incomplete), adding DOI to many papers
+
+* 2018-09-13
+ASIACCS 2018
+
+* 2018-09-11
+ToSC 2018 (Issue 3)
+
+* 2018-09-10
+WISA 2017, ACISP 2018, ACNS 2018, CRYPTO 2018, ESORICS 2018, ISC 2018, IWSEC 2018, JC 2018 (Number 3), NDSS 2018, PODC 2018, IEEE SP 2018, SCN 2018, STOC 2018, TRUSTBUS 2018
+
+* 2018-08-12
+EPRINT (up to 2018/733)
+
+* 2018-06-05
+ToSC 2018 (Issue 2)
+
+* 2018-05-08
+TCHES 2018 (Issue 2)
+
+* 2018-04-30
+AFRICACRYPT 2018, EUROCRYPT 2018, ICICS 2017, LATIN 2018, RSA 2018 
+
+* 2018-04-06
+FC 2017, FCW 2017, ICITS 2017, IMACC 2017, INDOCRYPT 2017, ISC 2017, ITCS 2017, ITCS 2018, NDSS 2017, SAC 2017, PKC 2018, SODA 2018, JC 2018 (Numbers 1-2)
+
+* 2018-03-05
+TCHES 2018 (Issue 1 and abbreviations)
+
+* 2018-02-27
+ToSC 2018 (Issue 1)
+
