+++ documentation/pandoc/macros.tex
@@ -1,0 +1,44 @@
+% alphabets
+\newcommand{\bC}{\overline{C}}
+\newcommand{\bM}{\mathsf{M}}
+\newcommand{\bS}{\mathsf{S}}
+\newcommand{\bm}{\mathsf{m}}
+
+\newcommand{\cA}{\mathcal{A}}
+\newcommand{\cB}{\mathcal{B}}
+\newcommand{\cC}{\mathcal{C}}
+\newcommand{\cE}{\mathcal{E}}
+\newcommand{\cD}{\mathcal{D}}
+\newcommand{\cG}{\mathcal{G}}
+\newcommand{\cI}{\mathcal{I}}
+\newcommand{\cH}{\mathcal{H}}
+\newcommand{\cJ}{\mathcal{J}}
+\newcommand{\cL}{\mathcal{L}}
+\newcommand{\cM}{\mathcal{M}}
+\newcommand{\cN}{\mathcal{N}}
+\newcommand{\cO}{\mathcal{O}}
+\newcommand{\cP}{\mathcal{P}}
+\newcommand{\cR}{\mathcal{R}}
+\newcommand{\cS}{\mathcal{S}}
+\newcommand{\cT}{\mathcal{T}}
+\newcommand{\cU}{\mathcal{U}}
+\newcommand{\cV}{\mathcal{V}}
+\newcommand{\cX}{\mathcal{X}}
+
+\newcommand{\SSEEnc}{\mathsf{Enc}_\mathsf{SSE}}
+
+\newcommand{\SymEnc}{\mathsf{Enc}_\mathsf{Sym}}
+\newcommand{\SymDec}{\mathsf{Dec}_\mathsf{Sym}}
+
+\newcommand{\ABEEnc}{\mathsf{Enc}_\mathsf{ABE}}
+\newcommand{\ABEDec}{\mathsf{Dec}_\mathsf{ABE}}
+
+\newcommand{\PRF}{\mathsf{PRF}}
+
+\newcommand{\key}{\mathsf{key}}
+\newcommand{\nonce}{\mathsf{nonce}}
+\newcommand{\mpk}{\mathsf{mpk}}
+\newcommand{\msk}{\mathsf{msk}}
+\newcommand{\sk}{\mathsf{sk}}
+\newcommand{\Attr}{\mathsf{Attr}}
+\newcommand{\UID}{\mathsf{UID}}
