+++ documentation/theme_overrides/partials/footer.html
@@ -1,0 +1,1 @@
+<footer class="md-footer"></footer>
